Added combat messages
